requirements.txt
openai==1.2.3  # Adjust to latest stable version if needed
requests==2.32.0
json5==0.9.14
pandas==2.2.2
numpy==1.26.4
openrouter==1.0.0
sendgrid~=6.11.0

Flask~=3.1.0
python-dotenv~=1.0.1
Markdown~=3.7
tabulate~=0.9.0